requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
pymilvus==2.4.0
alembic==1.13.1

# LLM & RAG
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.23
langgraph==0.0.20
openai==1.12.0

# Document Processing
pymupdf==1.23.21
python-multipart==0.0.6
python-docx==1.1.0
markdown==3.5.2

# Vector & Embedding
numpy==1.26.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.26.0
aiofiles==23.2.1
pandas==2.2.0
openpyxl==3.1.2

# Evaluation
rouge-score==0.1.2
sacrebleu==2.4.0
nltk==3.8.1

# RAG Techniques
rank-bm25==0.2.2
jieba==0.42.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2

